requirements.txt
streamlit
requests
beautifulsoup4
feedparser
pymongo
certifi
google-genai
toml